split off screensize
